# Add named pipes microbenchmark