`NamedPipeServerStream` instances can be reused. This PR adds a cache to the connection listener, allowing streams to be reused between connections.

A further improvement could be to completely reuse `NamedPipeConnection`. However, caching a connection context is harder because it has to determine whether anyone else could still use it. `NamedPipeServerStream` is private, making it relatively simple to detect whether it is still in use.

Before:
```
|    Method | ListenerQueueCount |     Mean |    Error |   StdDev |     Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |------------------- |---------:|---------:|---------:|---------:|------:|------:|------:|----------:|
| Plaintext |                  1 | 59.16 us | 1.130 us | 1.257 us | 16,902.9 |     - |     - |     - |     11 KB |
| Plaintext |                  2 | 59.62 us | 1.055 us | 0.935 us | 16,772.6 |     - |     - |     - |     11 KB |
| Plaintext |                  8 | 34.49 us | 0.508 us | 0.450 us | 28,997.5 |     - |     - |     - |     11 KB |
| Plaintext |                 16 | 35.30 us | 0.401 us | 0.335 us | 28,329.2 |     - |     - |     - |     11 KB |
```
After:
```
|    Method | ListenerQueueCount |     Mean |    Error |   StdDev |     Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |------------------- |---------:|---------:|---------:|---------:|------:|------:|------:|----------:|
| Plaintext |                  1 | 44.40 us | 0.644 us | 0.602 us | 22,522.2 |     - |     - |     - |      9 KB |
| Plaintext |                  2 | 29.90 us | 0.488 us | 0.433 us | 33,449.5 |     - |     - |     - |      9 KB |
| Plaintext |                  8 | 24.09 us | 0.219 us | 0.194 us | 41,509.1 |     - |     - |     - |      9 KB |
| Plaintext |                 16 | 23.98 us | 0.249 us | 0.221 us | 41,706.7 |     - |     - |     - |      9 KB |
```